feat: Add control to contract Api's

+++ business/src/main/java/org/grisbi/onefreelance/business/service/ContractService.java
@@ -2,10 +2,12 @@ package org.grisbi.onefreelance.business.service;
 
 import io.vavr.control.Try;
 import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.grisbi.onefreelance.business.mappers.ContractMapper;
+import org.grisbi.onefreelance.business.utils.UserUtils;
 import org.grisbi.onefreelance.model.dto.request.ContractRequest;
 import org.grisbi.onefreelance.model.dto.response.ContractResponse;
 import org.grisbi.onefreelance.model.errors.BusinessError;
@@ -22,6 +24,7 @@ import org.springframework.stereotype.Service;
 @RequiredArgsConstructor
 public class ContractService {
 
+  private static final String CONTRACT_NOT_FOUNT = "Contract Not Found";
   private final ContractRepository contractRepository;
   private final ContractMapper contractMapper;
 
@@ -32,8 +35,9 @@ public class ContractService {
    * @return contract response
    */
   public ContractResponse getContract(final UUID id) {
-    final ContractEntity contractEntity = contractRepository.findById(id)
-        .orElseThrow(() -> BusinessError.forError(ErrorHandler.NOT_FOUND, "Contract Not Found"));
+    final UUID connectedUser = UserUtils.getConnectedUser();
+    final ContractEntity contractEntity = contractRepository.findByIdAndCustomerId(id, connectedUser.toString())
+        .orElseThrow(() -> BusinessError.forError(ErrorHandler.NOT_FOUND, CONTRACT_NOT_FOUNT));
     return contractMapper.toContractResponse(contractEntity);
   }
 
@@ -43,10 +47,11 @@ public class ContractService {
    * @return contract response
    */
   public List<ContractResponse> getAllContracts() {
-    return contractRepository.findAll()
-        .stream()
-        .map(contractMapper::toContractResponse)
-        .toList();
+    return contractRepository.findAllByContractDataAndId(UserUtils.getConnectedUser().toString())
+        .map(contract -> contract.stream()
+            .map(contractMapper::toContractResponse)
+            .toList())
+        .orElseThrow(() -> BusinessError.forError(ErrorHandler.NOT_FOUND, CONTRACT_NOT_FOUNT));
   }
 
   /**
@@ -57,7 +62,7 @@ public class ContractService {
    */
   public ContractResponse createContract(final ContractRequest contractRequest) {
     final ContractEntity contractEntity = contractRepository
-        .save(contractMapper.toCreateContractEntity(contractRequest));
+        .save(contractMapper.toCreateContractEntity(contractRequest, UserUtils.getConnectedUser()));
     return contractMapper.toContractResponse(contractEntity);
   }
 
@@ -69,10 +74,17 @@ public class ContractService {
    * @return contract response
    */
   public ContractResponse updateContract(final UUID id, final ContractRequest contractRequest) {
-    final ContractEntity contractEntity = Try.of(() -> contractRepository
-            .save(contractMapper.toUpdateContractEntity(id, contractRequest)))
-        .getOrElseThrow(() -> BusinessError.forError(ErrorHandler.NOT_FOUND, "Contract Not Found"));
-    return contractMapper.toContractResponse(contractEntity);
+    final UUID connectedUser = UserUtils.getConnectedUser();
+    final Optional<ContractEntity> contract = contractRepository.findByIdAndCustomerId(id, connectedUser.toString());
+
+    if (contract.isPresent()) {
+      final ContractEntity contractEntity = Try.of(() -> contractRepository
+              .save(contractMapper.toUpdateContractEntity(id, contractRequest, connectedUser)))
+          .getOrElseThrow(() -> BusinessError.forError(ErrorHandler.NOT_FOUND, "Contract Not Found"));
+      return contractMapper.toContractResponse(contractEntity);
+    } else {
+      throw BusinessError.forError(ErrorHandler.NOT_FOUND, CONTRACT_NOT_FOUNT);
+    }
   }
 
   /**
@@ -81,6 +93,13 @@ public class ContractService {
    * @param id of contract
    */
   public void deleteContract(final UUID id) {
-    contractRepository.deleteById(id);
+
+    final Optional<ContractEntity> contractEntity = contractRepository.findByIdAndCustomerId(id,
+        UserUtils.getConnectedUser().toString());
+    if (contractEntity.isPresent()) {
+      contractRepository.deleteById(id);
+    } else {
+      throw BusinessError.forError(ErrorHandler.NOT_FOUND, CONTRACT_NOT_FOUNT);
+    }
   }
 }
